requirements.txt
click==8.0.3
dnspython==1.16.0
Flask==2.0.2
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pymongo==3.10.1
python-dotenv==0.19.2
typing-extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
